"Need o test admin part otherwise coding is done for admin part. Subscription is not included"

+++ app/src/main/java/com/test/skilllet/models/User.kt
@@ -9,7 +9,7 @@ class User() {
     var accType:String=""
     var phNumber:String=""
     var address: String =""
-    var rating:Int=0
+    var rating:Float=0.0f
     var key:String=""
     get() {
         return email.substring(0,email.indexOf("@"))
